Moved Observable and utils to a shared project
