+++ gui/gui.vcxproj
@@ -231,7 +231,7 @@
       <ConformanceMode>true</ConformanceMode>
       <PrecompiledHeader>Use</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\util</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\common</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp17</LanguageStandard>
     </ClCompile>
     <Link>
@@ -250,7 +250,7 @@
       <ConformanceMode>true</ConformanceMode>
       <PrecompiledHeader>Use</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\util</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\common</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp17</LanguageStandard>
       <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
     </ClCompile>
@@ -270,7 +270,7 @@
       <ConformanceMode>true</ConformanceMode>
       <PrecompiledHeader>Use</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\util</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\common</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp17</LanguageStandard>
     </ClCompile>
     <Link>
@@ -289,7 +289,7 @@
       <ConformanceMode>true</ConformanceMode>
       <PrecompiledHeader>Use</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\util</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>imgui\; imgui\backends;imgui\misc\cpp\;..\math\src; ..\math\eigen;..\common</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp17</LanguageStandard>
       <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
     </ClCompile>
